Update: HPOptim.lua, clean_up.sh, sample.lua, and README

+++ HPOptim/clean_up.sh
@@ -1,6 +1,21 @@
 # Author: Julien Hoachuck
 # Copyright 2015, Julien Hoachuck, All rights reserved.
-rm /data/db/spearmint.*
-rm /data/db/mongod.lock
+#!/bin/bash
+file1="0"
+file2="0"
+
+find . -maxdepth 1 -type f -name "spearmint.*" 2>/dev/null | grep -q . && file1="1"
+if [ $file1 -eq 1 ]; then
+	rm /data/db/spearmint.*
+	echo "Removed spearmint.*"
+fi
+
+find . -maxdepth 1 -type f -name "mongod.lock" 2>/dev/null | grep -q . && file2="1"
+if [ $file2 -eq 1 ]; then
+	rm /data/db/mongod.lock
+	echo "Removed mongod.lock"
+fi
+
 killall mongod
-mongod --fork --logpath /home/toor/Desktop/log --dbpath /data/db
+mongod
+
